requirements.txt
pandas
paho-mqtt
fastapi
sqlalchemy
psycopg2
numpy
typer